apps/ai-service/requirements.txt
fastapi==0.109.2
uvicorn==0.27.1
python-multipart==0.0.9
pydantic==2.6.1
pydantic-settings==2.1.0
google-cloud-vision==3.7.2
boto3==1.34.44
opencv-python==4.9.0.80
numpy==1.26.4
requests==2.31.0
python-dotenv==1.0.1
